# refactor(pkg): speed up opam file parsing